test(e2e): Update tests due to 0.17.0 changes (#2401)
* test(e2e): Only clean up org if org hasn't been deleted yet

* test(e2e): Use new worker filter tab

* test(e2e): Update selector for updated Manage button

Some pages now use a Manage button instead of a Manage link. The "Title" selector was not being picked up on the Manage button.

* test(e2e): Fix flakiness due to race condition

The test should navigate to the home page after creating all of the resources using the API

* test(e2e): Add cleanup to test

+++ ui/admin/tests/e2e/tests/delete-resources.spec.js
@@ -71,6 +71,7 @@ test.beforeEach(async () => {
 
 test('Verify resources can be deleted @ce @aws', async ({ page }) => {
   let orgId;
+  let orgDeleted = false;
   try {
     // Create boundary resources using CLI
     orgId = await createOrgCli();
@@ -191,9 +192,10 @@ test('Verify resources can be deleted @ce @aws', async ({ page }) => {
     // Delete org
     await page.goto(`/scopes/${orgId}/edit`);
     await deleteResource(page);
+    orgDeleted = true;
   } finally {
     // Delete org in case the test failed before deleting the org using UI
-    if (orgId) {
+    if (orgId && orgDeleted === false) {
       await deleteOrgCli(orgId);
     }
   }
